Conflicts:
	ConsomiTounsi/pom.xml
	ConsomiTounsi/src/main/java/tn/esprit/spring/Configuration/PaypalConfig.java
	ConsomiTounsi/src/main/java/tn/esprit/spring/Service/PaypalService.java
	ConsomiTounsi/src/main/java/tn/esprit/spring/controller/PaypalController.java
	ConsomiTounsi/src/main/java/tn/esprit/spring/entities/Command.java
	ConsomiTounsi/src/main/java/tn/esprit/spring/entities/Product.java
	ConsomiTounsi/src/main/java/tn/esprit/spring/entities/User.java
	ConsomiTounsi/src/main/resources/application.properties

+++ ConsomiTounsi/src/main/java/tn/esprit/spring/Repository/DeliveryMenRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import tn.esprit.spring.entities.DeliveryMen;
 
 @RepositoryRestResource(path = "DeliveryMen")
-public interface DeliveryMenRepository extends JpaRepository<DeliveryMen, Integer> {
+public interface DeliveryMenRepository extends JpaRepository<DeliveryMen, Long> {
 
 }
