+++ ConsomiTounsi/src/main/java/tn/esprit/spring/Service/CategoryServiceImpl.java
@@ -22,7 +22,7 @@ public class CategoryServiceImpl implements ICategoryService {
 	}
 
 	@Override
-	public void deleteCategory(int i) {
+	public void deleteCategory(Long i) {
 		CategoryRepository.deleteById(i);
 		
 	}
