+++ ConsomiTounsi/src/main/java/tn/esprit/spring/Service/UserServiceImpl.java
@@ -50,7 +50,7 @@ public class UserServiceImpl implements IUserService{
 	}
 
 	@Override
-	public void deleteUser(int id) {
+	public void deleteUser(Long id) {
 		userRepository.deleteById(id);
 		
 	}
@@ -62,7 +62,7 @@ public class UserServiceImpl implements IUserService{
 	}
 
 	@Override
-	public User retrieveUser(int id) {
+	public User retrieveUser(Long id) {
 		return userRepository.findById(id).get();
 	}
